moving file home back to features.
